requirements.txt
colorama==0.4.6
iniconfig==2.0.0
mysql-connector-python==8.3.0
packaging==23.2
pluggy==1.3.0
pytest==7.4.4
python-dotenv==1.0.1